Disconnect with an error message when receiving errors in early frames
This way it will only attempt to reconnect silently once, but if that
fails it won't loop forever, and instead it will show a hopefully useful
error message. Looking forward for new kinds of bug reports that aren't
"nothing works and i have no idea what's going on". Sorry for that.