Get rid of the endless SyncBatch spam
The initial batch of events sent several items in data, but we only
increased ack by one. The server expects us to increase ack by the
amount of items processed. So we were always behind, and the server kept
sending batch garbage, probably thinking we lost some packets.

It's stupid how it's super strict about some ways to drop packets but
tolerates and tries to recover some others. Or maybe it's not stupid and
I don't really get it yet.